:hammer: hoist shared i18n primitives to commonMain
Extract two pieces every platform (Desktop, Android, iOS) needs to keep in lockstep:

- SupportedLanguages: 10 language code constants + LANGUAGE_LIST. Adding a
  language here is a single-file edit visible to every platform copywriter,
  removing the drift that previously caused Android to miss `pt`.
- LanguageCache: a `(String) -> Copywriter` factory wrapped around
  io.ktor.util.collections.ConcurrentMap.computeIfAbsent. Replaces three
  near-identical caches (Desktop's ConcurrentHashMap, iOS's ktor ConcurrentMap,
  Android's separate AndroidCopywriterFactory class).

Desktop migration: DesktopGlobalCopywriter.Companion drops LANGUAGE_MAP and
the language constants — companion now holds only EMPTY_STRING. Cache moves
from process-static (companion) to instance-scoped (DI singleton), behaviorally
identical in production. Five call sites (I18n.kt, DesktopOCRModule.kt,
DesktopLocaleUtils.kt, two tests) now import constants from SupportedLanguages.

Mobile follow-up will live in the mobile repos.

Closes #4335
