+++ app/src/desktopMain/kotlin/com/crosspaste/module/ocr/DesktopOCRModule.kt
@@ -4,17 +4,17 @@ import com.crosspaste.app.AppFileType
 import com.crosspaste.app.DesktopAppWindowManager
 import com.crosspaste.app.WindowTrigger
 import com.crosspaste.config.DesktopConfigManager
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.DE
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.EN
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.ES
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.FA
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.FR
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.JA
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.KO
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.PT
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.ZH
-import com.crosspaste.i18n.DesktopGlobalCopywriter.Companion.ZH_HANT
 import com.crosspaste.i18n.GlobalCopywriter
+import com.crosspaste.i18n.SupportedLanguages.DE
+import com.crosspaste.i18n.SupportedLanguages.EN
+import com.crosspaste.i18n.SupportedLanguages.ES
+import com.crosspaste.i18n.SupportedLanguages.FA
+import com.crosspaste.i18n.SupportedLanguages.FR
+import com.crosspaste.i18n.SupportedLanguages.JA
+import com.crosspaste.i18n.SupportedLanguages.KO
+import com.crosspaste.i18n.SupportedLanguages.PT
+import com.crosspaste.i18n.SupportedLanguages.ZH
+import com.crosspaste.i18n.SupportedLanguages.ZH_HANT
 import com.crosspaste.image.OCRModule
 import com.crosspaste.module.DownloadState
 import com.crosspaste.module.DownloadTask
